feat: implement DenseTensorImpl with builder pattern and update related tests
