+++ orteaf/include/orteaf/internal/tensor/registry/tensor_impl_registry.h
@@ -14,18 +14,9 @@
 #include <orteaf/internal/storage/registry/storage_types.h>
 #include <orteaf/internal/tensor/manager/tensor_impl_manager.h>
 #include <orteaf/internal/tensor/manager/tensor_impl_manager.inl>
+#include <orteaf/internal/tensor/traits/tensor_impl_traits.h>
 
 namespace orteaf::internal::tensor::registry {
-
-// =============================================================================
-// TensorImplTraits
-// =============================================================================
-
-template <typename Impl> struct TensorImplTraits {
-  using Manager = TensorImplManager<Impl>;
-  using Lease = typename Manager::TensorImplLease;
-  static constexpr const char *name = "unknown";
-};
 
 // =============================================================================
 // TensorImplRegistry
@@ -34,23 +25,28 @@ template <typename Impl> struct TensorImplTraits {
 template <typename... Impls> class TensorImplRegistry {
 public:
   using StorageRegistry = ::orteaf::internal::storage::RegisteredStorages;
-  using ManagerTuple = std::tuple<typename TensorImplTraits<Impls>::Manager...>;
+  using ManagerTuple = std::tuple<::orteaf::internal::tensor::TensorImplManager<
+      Impls>...>;
 
-  using LeaseVariant =
-      std::variant<std::monostate, typename TensorImplTraits<Impls>::Lease...>;
+  using LeaseVariant = std::variant<std::monostate,
+                                    typename ::orteaf::internal::tensor::
+                                        TensorImplManager<Impls>::TensorImplLease...>;
 
   using ImplTypes = std::tuple<Impls...>;
 
   struct Config {
-    std::tuple<typename TensorImplTraits<Impls>::Manager::Config...> configs{};
+    std::tuple<typename ::orteaf::internal::tensor::TensorImplManager<
+        Impls>::Config...> configs{};
 
     template <typename Impl> auto &get() {
-      return std::get<typename TensorImplTraits<Impl>::Manager::Config>(
+      return std::get<typename ::orteaf::internal::tensor::TensorImplManager<
+          Impl>::Config>(
           configs);
     }
 
     template <typename Impl> const auto &get() const {
-      return std::get<typename TensorImplTraits<Impl>::Manager::Config>(
+      return std::get<typename ::orteaf::internal::tensor::TensorImplManager<
+          Impl>::Config>(
           configs);
     }
   };
@@ -70,11 +66,13 @@ public:
   bool isConfigured() const noexcept { return isConfiguredImpl<Impls...>(); }
 
   template <typename Impl> auto &get() {
-    return std::get<typename TensorImplTraits<Impl>::Manager>(managers_);
+    return std::get<::orteaf::internal::tensor::TensorImplManager<Impl>>(
+        managers_);
   }
 
   template <typename Impl> const auto &get() const {
-    return std::get<typename TensorImplTraits<Impl>::Manager>(managers_);
+    return std::get<::orteaf::internal::tensor::TensorImplManager<Impl>>(
+        managers_);
   }
 
   /// @brief Dispatch an operation to the correct manager based on lease type.
@@ -98,17 +96,19 @@ public:
 private:
   template <typename First, typename... Rest>
   void configureImpl(const Config &config, StorageRegistry &storage_registry) {
-    std::get<typename TensorImplTraits<First>::Manager>(managers_).configure(
-        std::get<typename TensorImplTraits<First>::Manager::Config>(
-            config.configs),
-        storage_registry);
+    std::get<::orteaf::internal::tensor::TensorImplManager<First>>(managers_)
+        .configure(
+            std::get<typename ::orteaf::internal::tensor::TensorImplManager<
+                First>::Config>(config.configs),
+            storage_registry);
     if constexpr (sizeof...(Rest) > 0) {
       configureImpl<Rest...>(config, storage_registry);
     }
   }
 
   template <typename First, typename... Rest> void shutdownImpl() {
-    std::get<typename TensorImplTraits<First>::Manager>(managers_).shutdown();
+    std::get<::orteaf::internal::tensor::TensorImplManager<First>>(managers_)
+        .shutdown();
     if constexpr (sizeof...(Rest) > 0) {
       shutdownImpl<Rest...>();
     }
@@ -116,8 +116,10 @@ private:
 
   template <typename First, typename... Rest>
   bool isConfiguredImpl() const noexcept {
-    bool first = std::get<typename TensorImplTraits<First>::Manager>(managers_)
-                     .isConfigured();
+    bool first =
+        std::get<::orteaf::internal::tensor::TensorImplManager<First>>(
+            managers_)
+            .isConfigured();
     if constexpr (sizeof...(Rest) > 0) {
       return first && isConfiguredImpl<Rest...>();
     }
@@ -127,8 +129,9 @@ private:
   // Dispatch by lease type
   template <typename Lease, typename First, typename... Rest, typename Func>
   static auto dispatchImpl(const Lease &lease, Func &&func) {
-    if constexpr (std::is_same_v<Lease,
-                                 typename TensorImplTraits<First>::Lease>) {
+    if constexpr (std::is_same_v<
+                      Lease, typename ::orteaf::internal::tensor::
+                                 TensorImplManager<First>::TensorImplLease>) {
       return func.template operator()<First>(lease);
     } else if constexpr (sizeof...(Rest) > 0) {
       return dispatchImpl<Lease, Rest...>(lease, std::forward<Func>(func));
@@ -140,7 +143,8 @@ private:
   bool dispatchByNameImpl(std::string_view name, Func &&func) {
     if (name == TensorImplTraits<First>::name) {
       func.template operator()<First>(
-          std::get<typename TensorImplTraits<First>::Manager>(managers_));
+          std::get<::orteaf::internal::tensor::TensorImplManager<First>>(
+              managers_));
       return true;
     }
     if constexpr (sizeof...(Rest) > 0) {
